+++ iris/service/content/file/document.py
@@ -4,10 +4,15 @@ from lovely.esdb.properties import Property
 from iris.service.db.dc import dc_defaults, DC_CREATED, DC_MODIFIED
 
 
-class File(Document):
-    """A file in the database
+class StorageType(object):
+    S3 = "s3"
 
-    TODO: needs to be completed
+
+class File(Document):
+    """A file in the database.
+
+    This is a meta data container for the file. The file remains in an external
+    storage.
     """
 
     INDEX = 'files'
@@ -18,3 +23,13 @@ class File(Document):
         default=dc_defaults(DC_CREATED, DC_MODIFIED),
         doc="Dublin Core data."
     )
+
+    state = Property()
+
+    original_name = Property()
+
+    owner_id = Property()
+
+    storage_type = Property()
+
+    content_type = Property()